basic start

+++ ElectricInterface.java
@@ -2,8 +2,8 @@
 public interface ElectricInterface {
 
     double calcMPGe();
-    void setElectricMiles(int totalElectricMiles);
-    void setTotalkWh(int totalkWh);
+    void setElectricMiles(double totalElectricMiles);
+    void setTotalkWh(double totalkWh);
 
 }
